requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
pymongo[srv]==4.3.3
python-dotenv==0.19.0
gunicorn==20.1.0
flask-cors==3.0.10
dnspython==2.4.2
Werkzeug==2.0.1
marshmallow==3.19.0
python-dateutil==2.8.2